Aula 04 de Flutter + Parte do trabalho

+++ 06-Flutter-Marcos/Trabalho/lib/datasources/elemento_datasource.dart
@@ -2,59 +2,69 @@ import 'package:sqflite/sqflite.dart';
 import 'package:trabalho/datasources/banco_dados.dart';
 import 'package:trabalho/models/models.dart';
 
-class ElementoDatasource {
-  Future<void> insert(String tabela, Elemento elemento) async {
+abstract class ElementoDatasource<T extends Elemento> {
+  final String nomeTabela;
+
+  ElementoDatasource(this.nomeTabela);
+
+  String createTableSql();
+
+  Future<void> insert(T elemento) async {
     Database db = await BancoDados().database;
 
     await db.insert(
-      tabela,
+      nomeTabela,
       elemento.toMap(),
       conflictAlgorithm: ConflictAlgorithm.replace,
     );
   }
 
-  Future<void> update(String tabela, Elemento elemento) async {
+  Future<void> update(Elemento elemento) async {
     Database db = await BancoDados().database;
 
     await db.update(
-      tabela,
+      nomeTabela,
       elemento.toMap(),
       where: 'id = ?',
       whereArgs: [elemento.id],
     );
   }
 
-  Future<void> delete(String tabela, String id) async {
+  Future<void> delete(String id) async {
     Database db = await BancoDados().database;
 
+    if (!podeDeletar(id)) {
+      throw Exception('Não é possível deletar este elemento da tabela $nomeTabela com id $id');
+    }
+
     await db.delete(
-      tabela,
+      nomeTabela,
       where: 'id = ?',
       whereArgs: [id],
     );
   }
 
-  Future<List<Elemento>> getAll(String tabela, Elemento modelo) async {
+  Future<List<T>> getAll();
+  Future<List<Map<String, Object?>>> getAllDatasource() async {
     Database db = await BancoDados().database;
 
-    final List<Map<String, dynamic>> maps = await db.query(tabela);
-
-    return maps.map((elemento) => modelo.fromMap(elemento)).toList();
+    return await db.query(nomeTabela);
   }
 
-  Future<Elemento> getById(String tabela, Elemento modelo, String id) async {
+  Future<T> getById(String id);
+  Future<Map<String, dynamic>> getByIdDatasource(String id) async {
     Database db = await BancoDados().database;
 
     final List<Map<String, dynamic>> maps = await db.query(
-      tabela,
+      nomeTabela,
       where: 'id = ?',
       whereArgs: [id],
     );
 
-    return modelo.fromMap(maps.first);
+    return maps.first;
   }
-}
 
-abstract class ElementoDatasourceSql implements ElementoDatasource {
-  String createTable();
+  bool podeDeletar(String id) {
+    return true;
+  }
 }
